Fix handling of gamepad state
Chrome's Gamepad objects are only snapshots of the state at the moment the
object was obtained, rather than being a live object kept up-to-date with the
hardware state.

That is why scangamepads() has to be called on all iterations to obtain a new
copy of the Gamepad objects. However, that means we lose the "previous" state
tracking we added ourselves to detect changes.

This change fixes that by saving the previous state when obtaining the new
state snapshot.